video2mp3/requirements.txt
ffmpeg-python==0.2.0
google-cloud-storage==2.7.0